requirements.txt
langchain==0.3.15
langchain-community==0.3.15
langchain-core==0.3.31
langchain-openai==0.3.2
langchain-text-splitters==0.3.5

faiss-cpu==1.9.0.post1
openai==1.60.1

requests==2.32.5
beautifulsoup4==4.13.3

fastapi==0.118.0
uvicorn==0.37.0
pydantic==2.11.7
nest-asyncio==1.6.0

typing-extensions==4.12.2
python-dotenv==1.1.1